requirements.txt
oneflow~=0.8.0.dev20220630+cu112
torch~=1.11.0
numpy~=1.23.0
pycocotools~=2.0.4
Pillow~=9.1.1
torchvision~=0.12.0
matplotlib~=3.5.2
Flask~=2.1.2
opencv-python~=4.6.0.66
flowvision~=0.1.0